// frontend/Open2Work/src/interfaces/teamInterface.ts
// Generated by https://quicktype.io

import { User } from "./loginInterface";

export interface Team {
    _id:          string;
    devs:         User[];
    language:     string[];
    isComplete:   boolean;
    time_zone:    string[];
    working:      boolean;
    availability: string;
    __v:          number;
    stack:        string;
    orders: OrderElement[];
}

export interface OrderElement {
    _id:         string;
    client:      string;
    team:        string;
    devs_ok:     any[];
    devs_not:    any[];
    description: string;
    createdAt:   string;
    __v:         number;
}
// export interface Dev {
//     _id:         string;
//     name:        string;
//     surname:     string;
//     email:       string;
//     password:    string;
//     role:        string;
//     avatar:      string;
//     isDev:       boolean;
//     social:      Social;
//     info:        Info;
//     oldTeams:    any[];
//     __v:         number;
//     currentTeam: string;
// }

// export interface Info {
//     time_availability: string;
//     time_zone:         string;
//     experience:        number;
//     language:          string[];
//     _id:               string;
// }

// export interface Social {
//     linkedin:  string;
//     portfolio: string;
//     github:    string;
//     _id:       string;
// }